git: configure allowedSignersFile
Referenced file does not need to have any content, it is just to allow
the use of SSH keys for commit signature. Or something. Boils down to
the common pattern of "Security cannot just work, it has to be fiddled
with until you don't know why it (seemingly) works." Security is cringe.